+++ crates/ruma-client-api/src/session.rs
@@ -5,5 +5,7 @@ pub mod login;
 pub mod login_fallback;
 pub mod logout;
 pub mod logout_all;
+#[cfg(feature = "unstable-msc2918")]
+pub mod refresh_token;
 pub mod sso_login;
 pub mod sso_login_with_provider;