+++ crates/ruma-client-api/src/session/login.rs
@@ -5,6 +5,9 @@ pub mod v3 {
     //!
     //! [spec]: https://spec.matrix.org/v1.2/client-server-api/#post_matrixclientv3login
 
+    #[cfg(feature = "unstable-msc2918")]
+    use std::time::Duration;
+
     use ruma_common::{
         api::ruma_api,
         serde::{Incoming, JsonObject},
@@ -44,6 +47,16 @@ pub mod v3 {
             /// Ignored if `device_id` corresponds to a known device.
             #[serde(skip_serializing_if = "Option::is_none")]
             pub initial_device_display_name: Option<&'a str>,
+
+            /// If set to `true`, the client supports refresh tokens.
+            #[cfg(feature = "unstable-msc2918")]
+            #[serde(
+                default,
+                skip_serializing_if = "ruma_common::serde::is_default",
+                rename = "org.matrix.msc2918.refresh_token",
+                alias = "refresh_token",
+            )]
+            pub refresh_token: bool,
         }
 
         response: {
@@ -71,6 +84,30 @@ pub mod v3 {
             /// If present, clients SHOULD use the provided object to reconfigure themselves.
             #[serde(skip_serializing_if = "Option::is_none")]
             pub well_known: Option<DiscoveryInfo>,
+
+            /// A refresh token for the account.
+            ///
+            /// This token can be used to obtain a new access token when it expires by calling the
+            /// `/refresh` endpoint.
+            #[cfg(feature = "unstable-msc2918")]
+            #[serde(skip_serializing_if = "Option::is_none")]
+            pub refresh_token: Option<String>,
+
+            /// The lifetime of the access token, in milliseconds.
+            ///
+            /// Once the access token has expired, a new access token can be obtained by using the
+            /// provided refresh token. If no refresh token is provided, the client will need to
+            /// re-login to obtain a new access token.
+            ///
+            /// If this is `None`, the client can assume that the access token will not expire.
+            #[cfg(feature = "unstable-msc2918")]
+            #[serde(
+                with = "ruma_common::serde::duration::opt_ms",
+                default,
+                skip_serializing_if = "Option::is_none",
+                rename = "expires_in_ms",
+            )]
+            pub expires_in: Option<Duration>,
         }
 
         error: crate::Error
@@ -79,14 +116,30 @@ pub mod v3 {
     impl<'a> Request<'a> {
         /// Creates a new `Request` with the given login info.
         pub fn new(login_info: LoginInfo<'a>) -> Self {
-            Self { login_info, device_id: None, initial_device_display_name: None }
+            Self {
+                login_info,
+                device_id: None,
+                initial_device_display_name: None,
+                #[cfg(feature = "unstable-msc2918")]
+                refresh_token: false,
+            }
         }
     }
 
     impl Response {
         /// Creates a new `Response` with the given user ID, access token and device ID.
         pub fn new(user_id: OwnedUserId, access_token: String, device_id: OwnedDeviceId) -> Self {
-            Self { user_id, access_token, home_server: None, device_id, well_known: None }
+            Self {
+                user_id,
+                access_token,
+                home_server: None,
+                device_id,
+                well_known: None,
+                #[cfg(feature = "unstable-msc2918")]
+                refresh_token: None,
+                #[cfg(feature = "unstable-msc2918")]
+                expires_in: None,
+            }
         }
     }
 
@@ -374,6 +427,8 @@ pub mod v3 {
                 login_info: LoginInfo::Token(Token { token: "0xdeadbeef" }),
                 device_id: None,
                 initial_device_display_name: Some("test"),
+                #[cfg(feature = "unstable-msc2918")]
+                refresh_token: false,
             }
             .try_into_http_request(
                 "https://homeserver.tld",
@@ -402,6 +457,8 @@ pub mod v3 {
                 }),
                 device_id: None,
                 initial_device_display_name: Some("test"),
+                #[cfg(feature = "unstable-msc2918")]
+                refresh_token: false,
             }
             .try_into_http_request(
                 "https://homeserver.tld",