+++ crates/ruma-client-api/CHANGELOG.md
@@ -5,6 +5,10 @@ Breaking changes:
 * Remove `PartialEq` implementations for a number of types
   * If the lack of such an `impl` causes problems, please open a GitHub issue
 
+Improvements:
+
+* Add unstable support for refresh tokens (MSC2918)
+
 # 0.14.1
 
 Improvements: